Fixed completely broken Thai date formatting unit tests, and then fixed the code to pass those tests properly. Thai solar calendar works properly now.
Added "calendar" option to the date formatter factory method in addition to "type". The two mean the same thing.

Added some tests to make sure that formatting a date in any calendar works in any locale.

Fixed the date range formats for gregorian in Iran. The problem was that the date range object was not looking at the calendar option, even though the inline docs said it would.

Both the Persian astronomical and the Persian algorithm calendars work fully now and pass all unit tests. The Persian algorithmic calendar is named "persian" and the algorithmic one is "persian-algo". All unit tests pass. Add "persian" and "persian-algo" to the list of calendars in the default date formats. Switched Iran and Afghanistan to use the new persian calendar.

Fixed up date formats for Kurdish, Hebrew, Pashto, and Urdu to appear correct in RTL.