Implemented getAncestorTranslationUnit()
Change getTranslationUnit so that if it cannot find an exact match for the locale, it returns a compatible translation instead. (ie. a request for en-US will return the string for en-Latn-US)

Fixed the tests for getAncestorTranslationUnit so that the unit returned has the exact locale of the unit rather than the one requested. The locale of the unit may be compatible instead of exactly equal to the requested locale.